move serve-static-assets file to lib
