+++ server/utils/serve-static-assets.js
